Add state transitions to TPX Jam
Added all state transititions from the FPGA/BP I can think of. Built
in logic for transisitions, as well as telling the app about what is
going on.

Made some differences in the topics. Trying to switch between topics
for anything related to a jam (leaving, joining, etc), the teleporters
telling the apps what is going on, and the teleporter being able to
tell the platform about something (as of yet undetermined). The reason
for adding platform, is that I saw we immediately subscribed to all
kinds of topics from the platform, but we really don't need that much
traffic.