Add Jam Platform
Jam Platform handles Teleporters asking for a jam. When two
Teleporters have asked for a jam, a Jam process starts where each
Teleporter is told to join a jam and the App is told about the Jam.

A timeout is present for an ask. Default is 5 minutes. If no other
Teleporter has asked as well during that time, the Teleporter leaves
the ask and is now free.