Add graceful shutdown of a jam
There are a couple of things that can go wrong when a jam is stopped,
including a teleporter not actually leaving. This is a first attempt
at fixing this state, where the jam session management tries to be
smart about how things are shut down.

This has the added benefit of showing the App different states when
shutting down.