Add local-ip/public-ip and clean up tests
The move from :jam to :jams in the platform db messed up the tests, as
they were forgotten.

Added local-ip/public-ip to the messages. Not enforced.

Re-enabled the handle-ipc in TPX Jam, while removing everything SIP
related.

Simplified the jam topics to only one. This is meant to be used by the
app alone. It's used for updates from a jam during a session.